+++ otterdog/operations/diff_operation.py
@@ -268,8 +268,7 @@ class DiffOperation(Operation):
             modified_repo = expected_repo.get_difference_from(current_repo)
 
             if len(modified_repo) > 0:
-                self.handle_modified_repo(github_id, current_repo_name, modified_repo)
-                diff_status.differences += len(modified_repo)
+                diff_status.differences = self.handle_modified_repo(github_id, current_repo_name, modified_repo)
 
             self._process_branch_protection_rules(github_id,
                                                   current_repo,
@@ -367,7 +366,7 @@ class DiffOperation(Operation):
     def handle_modified_repo(self,
                              org_id: str,
                              repo_name: str,
-                             modified_repo: dict[str, Change[Any]]) -> None:
+                             modified_repo: dict[str, Change[Any]]) -> int:
         raise NotImplementedError
 
     @abstractmethod